requirements.txt
Django==4.2
djangorestframework==3.14.0
pydantic==1.10.7
mysqlclient==2.2.4
python-dateutil>=2.8.2
python-dotenv==1.0.0
